refactor(bootstrap): A66 step 6 - split phases 5-10 into per-group files + cleanup
New files:
- phases_lifecycle.go (147 lines): phase3b LifecycleContributor discovery +
  registerAllHealthCheckers / registerCellHealthCheckers / registerOneCellHealthCheckers /
  registerConfigDriftChecker
  ref: uber-go/fx lifecycle.go
- phases_http.go (430 lines): phase5 router construction, route group mounting,
  auth plan wiring + validation helpers
  ref: kubernetes/kubernetes apiserver/pkg/server/genericapiserver.go
- phases_events.go (110 lines): phase6 event router startup + checkNoEventRegistrars
  ref: uber-go/fx app.go (run vs stop ctx separation)
- phases_workers.go (67 lines): phase8 worker group startup
  ref: uber-go/fx app.go (background goroutine lifecycle coordination)
- phases_shutdown.go (190 lines): phases 9-10 signal wait + LIFO graceful shutdown
  ref: sigs.k8s.io/controller-runtime pkg/manager/internal.go (engageStopProcedure)

bootstrap_phases.go reduced from 954 → 123 lines (phaseState + helpers only).

Co-1 (dead code): removed WithDisableObservabilityRestore + disableObservabilityRestore
field (set but never read; phase6 always adds ContractTracingMiddleware unconditionally).
Removed corresponding test TestBootstrap_EventSubscriptions_DisableObservabilityRestore.

Te-7 (test style): migrated lifecycle_test.go from bare t.Fatal/t.Errorf to
testify require/assert, consistent with bootstrap_test.go and phases_test.go.
Removed now-unused equalStrSlice helper.

Op-9: SKIPPED — lifecycle.go hook.timeout log already carries hookIdentityAttrs(h)
which includes slog.String("name", h.Name); equivalent field present.

Regenerated assemblies/corebundle/generated/metrics-schema.yaml; http_request_duration_seconds
source file is now phases_http.go (was bootstrap_phases.go).
Updated tools/metricschema/schema_test.go to pin new file reference.
